# Shortcut in `appendNewNode`